Simplify logic, eliminate collapse flashes (#24)
The changes here vastly simplify the animation logic. Instead of using two effect hooks to stage synchronous changes, this reduces them down a single hook using nested RAF callbacks.

The bigger achievement here is finding the conditions in which the collapse will jump when beginning to transition the other direction at the end of the first transition. I've added a couple of conditionals that give the component the option to "bail out" of the end of transition style overrides that caused the collapse to jump.